# swagger.yaml
swagger: '2.0'
info:
  title: Process Service
  termsOfService: http://swagger.io/terms/
  description: 'Process Service is an abstract and limiting API for Camunda.'
  version: '1.0'
  contact:
    name: sh20xyqi
    email: sh20xyqi@studserv.uni-leipzig.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - https
host: api.sepl.infai.org
basePath: /process/service
produces:
  - application/json
  - text/plain
paths:
...
  /deployment:
    get:
      summary:
        Returns the list of deployments.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - in: header
          name: id
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description: The list of deployments.
          schema:
            type: array
            items:
              $ref: '#/definitions/Deployment'
        500:
          description: Server was unable to return the deployments.
          schema:
            type: string
            format: string
  /deployment/{id}:
    get:
      summary:
        Returns the deployment identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: id
          in: path
          description: The passed deployment id.
          required: true
          type: string
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the selected deployment.
          schema:
            $ref: '#/definitions/Deployment'
        500:
          description:
            Server was unable to return the deployment.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the deployment.
          schema:
            type: string
            format: string
  /deployment/{id}/definition:
    get:
      summary:
        Returns the process definition identified by its deployment id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: id
          in: path
          description: The passed deployment id.
          required: true
          type: string
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the specific process definition.
          schema:
            $ref: '#/definitions/ProcessDefinition'
        500:
          description:
            Server was unable to return the process definition.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the deployment.
          schema:
            type: string
            format: string
  /process-definition/{id}:
    get:
      summary:
        Returns the process definition identified by its own id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: id
          in: path
          description: The passed process id.
          required: true
          type: string
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the specific process definition.
          schema:
            $ref: '#/definitions/ProcessDefinition'
        500:
          description:
            Server was unable to return the process definition.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the deployment.
          schema:
            type: string
            format: string
  /process-definition/{id}/start:
    get:
      summary:
        Starts the process identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: id
          in: path
          description: The passed process id.
          required: true
          type: string
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns names and values of the started processes.
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessVariable'
        500:
          description:
            Server was unable the start the process.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the deployment.
          schema:
            type: string
            format: string
  /process-definition/{id}/diagram:
    get:
      summary:
        Returns the process definition diagram identified by its process id.
      description:
        ''
      produces:
        - application/json
        - image/svg+xml
        - text/plain
      parameters:
        - name: id
          in: path
          description: The passed process id.
          required: true
          type: string
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the process definition diagram.
          schema:
            type: file
            format: image/svg+xml
        500:
          description:
            Server was unable the start the process.
          schema:
            type: string
            format: string
        501:
          description:
            Server was unable to stream diagram to response body.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the process definition diagram.
          schema:
            type: string
            format: string
  /process-definition/{id}/incident/count:
    get:
      summary:
        Returns incidients relating to the process definition.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: id
          in: path
          description: The passed process id.
          required: true
          type: string
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns how many incidients occured relating to the passed process definition.
          schema:
            type: integer
            format: int64
        500:
          description:
            Server was unable the start the process.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the deployment.
          schema:
            type: string
            format: string
  /process-instance:
    get:
      summary:
        Returns a list of all process instances relating to the user id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            A list of process instances.
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessInstance'
        500:
          description:
            Server was unable to return the list.
          schema:
            type: string
            format: string
  /process-instances/count:
    get:
      summary:
        Returns the amount of processes instances relating to the user id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            The amount of processes instances.
          schema:
            type: integer
            format: int64
        500:
          description:
            Server was unable to count the process instances.
          schema:
            type: string
            format: string
  /process-instance/{id}:
    delete:
      summary:
        Deletes the process instance identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed instance id.
          required: true
          type: string
      responses:
        204:
          description:
            Returns that the deletion was successful.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permission to delete the process instance.
          schema:
            type: string
            format: string
        500:
          description:
            Server was unable to delete the process instance.
          schema:
            type: string
            format: string
  /process-instance/{id}/incident:
    get:
      summary:
        Returns a list of incidents relating to a specific process instance identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed instance id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the list of incidents of a process instance.
          schema:
            type: array
            items:
              $ref: '#/definitions/Incident'
        401:
          description:
            User has no permission to get the incidents.
          schema:
            type: string
            format: string
        500:
          description:
            Server was unable to return the list.
          schema:
            type: string
            format: string
  /history/process-instance:
    get:
      summary:
        Returns the history of all process instances belonging to the user identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the list of the history of selected process instances.
          schema:
            type: array
            items:
              $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the list.
          schema:
            type: string
            format: string
  /history/finished/process-instance:
    get:
      summary:
        Returns the history of all finished process instances belonging to the user identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the list of the history of selected finished process instances.
          schema:
            type: array
            items:
              $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the list.
          schema:
            type: string
            format: string
  /history/unfinished/process-instance:
    get:
      summary:
        Returns the list of all unfinished process instances belonging to the user identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the list of the history of selected unfinished process instances.
          schema:
            type: array
            items:
              $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the list.
          schema:
            type: string
            format: string
  /history/process-instance/{id}:
    delete:
      summary:
        Removes the history of a process instance identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed process instance id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the deleted history of the process instance.
          schema:
            $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the list.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to delete the history of the process instance.
          schema:
            type: string
            format: string
  /history/process-definition/{id}/process-instance:
    get:
      summary:
        Returns the history of the process instance identified by its process definition id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed process definition id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the history of the selected process instance.
          schema:
            $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the history.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the history of the process instance.
          schema:
            type: string
            format: string
  /history/process-definition/{id}/process-instance/finished:
    get:
      summary:
        Returns the history of the process instance identified by its process definition id whereby the process instance has to be finished.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed process definition id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the history of the selected process instance.
          schema:
            $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the history.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the history of the process instance.
          schema:
            type: string
            format: string
  /history/process-definition/{id}/process-instance/unfinished:
    get:
      summary:
        Returns the history of the process instance identified by its process definition id whereby the process instance has to be unfinished.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed process definition id.
          required: true
          type: string
      responses:
        200:
          description:
            Returns the history of the selected process instance.
          schema:
            $ref: '#/definitions/HistoricProcessInstance'
        500:
          description:
            Server was unable to return the history.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the history of the process instance.
          schema:
            type: string
            format: string
  /history/process-instance/{id}/variable-instance:
    get:
      summary:
        Returns a single historic data identified by its id.
      description:
        ''
      produces:
        - application/json
        - text/plain
      parameters:
        - name: user_id
          in: header
          description: The passed user id.
          required: true
          type: string
        - name: id
          in: path
          description: The passed variable instance id.
          required: true
          type: string
      responses:
        200:
          description:
            The single historic data of the process instance.
          schema:
            $ref: '#/definitions/VariableInstance'
        500:
          description:
            Server was unbale to return the historic data.
          schema:
            type: string
            format: string
        401:
          description:
            User has no permissions to get the historic data.
          schema:
            type: string
            format: string
definitions:
...
  Deployment:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      source:
        type: string
        format: string
      tenantId:
        type: string
        format: string
      deploymentTime:
        type: string
        format: datetime
    required:
      - id
      - name
      - source
      - tenantId
      - deploymentTime
  ProcessDefinition:
    type: object
    properties:
      id:
        type: string
        format: string
      key:
        type: string
        format: string
      category:
        type: string
        format: string
      description:
        type: string
        format: string
      name:
        type: string
        format: string
      version:
        type: number
        format: float
      resource:
        type: string
        format: string
      deploymentId:
        type: string
        format: string
      diagram:
        type: string
        format: string
      suspended:
        type: boolean
      tenantId:
        type: string
        format: string
      versionTag:
        type: string
        format: string
    required:
      - id
      - key
      - category
      - description
      - name
      - version
      - resource
      - deploymentId
      - diagram
      - suspended
      - tenantId
      - versionTag
  ProcessVariable:
    type: object
    properties:
      name:
        type: string
        format: string
      value:
        type: object
    required:
      - name
      - value
...
  ProcessInstance:
    type: object
    properties:
      id:
        type: string
        format: string
      definitionId:
        type: string
        format: string
      businessKey:
        type: string
        format: string
      caseInstanceId:
        type: string
        format: string
      ended:
        type: boolean
      suspended:
        type: boolean
      tenantId:
        type: string
        format: string